chore: remove separate recursive paths and unify
