+++ packages/core/admin/admin/src/content-manager/pages/NoContentTypePage.tsx
